Adjusted PriceTier.Price property to be a double

+++ src/shared/BP-POC.Domain/Printers/PriceTier.cs
@@ -4,9 +4,9 @@ public class PriceTier
 {
     public int Floor { get; private set; }
     public int Ceiling { get; private set; }
-    public decimal Price { get; private set; }
+    public double Price { get; private set; }
 
-    public PriceTier(int floor, int ceiling, decimal price)
+    public PriceTier(int floor, int ceiling, double price)
     {
         Floor = Guard.Against.Negative(floor, nameof(floor));
         Ceiling = Guard.Against.NegativeOrZero(ceiling, nameof(ceiling));
